pyproject.toml
[tool.poetry]
name = "flightapi"
version = "0.1.0"
description = "api for flight search"
authors = ["Franco Lema <francoe.lema@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.116.1"
uvicorn = "^0.35.0"
psycopg2-binary = "^2.9.10"
sqlalchemy = "^2.0.41"
alembic = "^1.16.4"
pydantic = "^2.11.7"
python-dotenv = "^1.1.1"
asyncpg = "^0.30.0"
python-dateutil = "^2.9.0.post0"
logger = "^1.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
